requirements.txt
bidict==0.21.2
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
python-engineio==3.14.2
python-socketio==4.6.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
WTForms==2.3.3
